backend/requirements.txt
flask
flask_sqlalchemy
pyjwt
passlib
flask_cors
